accessible/tests/mochitest/name/test_browserui.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Accessibility Name Calculating Test.">

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../states.js"></script>
  <script type="application/javascript"
          src="../events.js"></script>
  <script type="application/javascript"
          src="../browser.js"></script>

  <script type="application/javascript">
  <![CDATA[
    function addTab(aURL)
    {
      this.eventSeq = [
        new invokerChecker(EVENT_DOCUMENT_LOAD_COMPLETE, tabDocumentAt, 1)
      ];

      this.invoke = function addTab_invoke()
      {
        tabBrowser().addTrustedTab(aURL);
      }

      this.getID = function addTab_getID()
      {
        return "add tab: " + aURL;
      }
    }

    function switchTab(aTabBrowser, aWindow)
    {
      this.invoke = function switchTab_invoke()
      {
        synthesizeKey("VK_TAB", { ctrlKey: true }, browserWindow());
      }

      this.eventSeq = [
        new focusChecker(tabDocumentAt, 1)
      ];

      this.check = function switchTab_check(aEvent)
      {
        var title = getAccessible(browserDocument()).name;
        isnot(title.indexOf(aEvent.accessible.name), -1,
              "Window title contains the name of active tab document" +
              " (Is '" + aEvent.accessible.name + "' in '" + title + "'?)");
      }

      this.getID = function switchTab_getID() { return "switch tab"; }
    }

    ////////////////////////////////////////////////////////////////////////////
    // Tests

    //gA11yEventDumpID = "eventdump"; // debug stuff
    //gA11yEventDumpToConsole = true; // debug

    var gQueue = null;
    function doTests()
    {
      gQueue = new eventQueue();
      gQueue.push(new addTab("about:mozilla"));
      gQueue.push(new switchTab());
      gQueue.onFinish = function()
      {
        closeBrowserWindow();
      }

      gQueue.invoke();
    }

    SimpleTest.waitForExplicitFinish();
    openBrowserWindow(doTests, "about:license");
  ]]>
  </script>

  <vbox flex="1" style="overflow: auto;">
  <body xmlns="http://www.w3.org/1999/xhtml">
    <a target="_blank"
       href="https://bugzilla.mozilla.org/show_bug.cgi?id=507382"
       title="focus is fired earlier than root accessible name is changed when switching between tabs">
      Mozilla Bug 
    </a>
    <p id="display"></p>
    <div id="content" style="display: none">
    </div>
    <pre id="test">
    </pre>
  </body>

  <vbox id="eventdump"></vbox>
  </vbox>
</window>
